Add new image assets and update .gitignore
Introduce new image files in various formats to the public/assets/images directory to enhance the visual content. This includes banner and logo images in png, webp, and ico formats, along with a new SVG placeholder. Additionally, update the .gitignore file to allow asset tracking in version control by removing the exclusion of the 'assets/' directory.